Mini cambio de int a float

+++ Practica1/AndroidEngine/src/main/java/com/example/aengine/AGraphics.java
@@ -81,12 +81,12 @@ public class AGraphics implements IGraphics {
 
         int w,h;
         if (this.scaleInX){
-            this.offsetX = (int)((getWidth() / 2.0f) - (((float) this.logicWidth / 2.0f) * this.scaleFactor));
+            this.offsetX = (int)(((float)getWidth() / 2.0f) - (((float) this.logicWidth / 2.0f) * this.scaleFactor));
             this.offsetY = 0;
         }
         else{
             this.offsetX = 0;
-            this.offsetY = (int)((getHeight() / 2.0f) - (((float) this.logicHeight / 2.0f)  * this.scaleFactor));
+            this.offsetY = (int)(((float)getHeight() / 2.0f) - (((float) this.logicHeight / 2.0f)  * this.scaleFactor));
         }
         w = (int)((float)this.logicWidth * this.scaleFactor);
         h = (int)((float)this.logicHeight * this.scaleFactor);
